refactor(server): resolve remaining groupBy v2 review comments
- Extract buildGroupByRecordConditions into a shared util used by both the v1
  and v2 with-records services (single source of truth for the group-scoping
  WHERE), removing the duplicated private methods.
- Drop the as-unknown-as RecordQueryBuilder cast: the v2 builder satisfies the
  structural RecordQueryBuilder type directly.
